requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.6
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
